Crate cubecl_runtime

CubeCL runtime crate that helps creating high performance async runtimes.

Modules

benchmark

Module for benchmarking any executable part

channel

Compute channel module.

client

Compute client module.

debug

Debugging utilities.

memory_management

Memory management module.

server

Compute server module.

storage

Compute Storage module.

tune

Autotune module

Macros

local_tuner

Create a local tuner with the provided name.

memory_id_type

Create new memory ID types.

storage_id_type

Create a new storage ID type.

tune_with

Tune the operation set with these benchmark inputs

Structs

ComputeRuntime

The compute type has the responsibility to retrieve the correct compute client based on the given device.

DeviceProperties

Properties of what the device can do, like what features are supported by it and what its memory properties are.

Enums

ExecutionMode

The kind of execution to be performed.

TimestampsError

Error that can occurred when collecting timestamps from a device.

Type Aliases

TimestampsResult

Result when collecting timestamps.