cubecl_runtime

Module server

Compute server module.

Structs

Binding

Binding of a tensor handle to execute a kernel.

Handle

Server handle containing the memory handle.

Enums

CubeCount

Specifieds the number of cubes to be dispatched for a kernel.

Traits

ComputeServer

The compute server is responsible for handling resources and computations over resources.