cubecl_runtime

Module tune

Autotune module

Macros

local_tuner

Create a local tuner with the provided name.

tune_with

Tune the operation set with these benchmark inputs

Structs

LocalTuner

A local tuner allows to create a tuner for a specific key that can be different from the server key.

TuneBenchmark

A benchmark that runs on server handles

Tuner

Executes autotune benchmarking and caching

Enums

AutotuneError

Error from running autotune.

TuneCacheResult

Result of the cache try

Traits

AutotuneKey

Trait alias with support for persistent caching

AutotuneOperation

Contains operation to run and inputs on which to run it

AutotuneOperationSet

Groups operations of the same type for autotune

Functions

anchor

Anchor a number to a power of 2.

compute_checksum

Default checksum for an operation set

get_persistent_cache_file_path

Return the file path for the persistent cache on disk prefix should be the device id computed at the backend level